[FIX] web_studio: fix bugs in edition of kanban
Before this commit there were two issues:
- add a menu on a card
- try to remove or add a field elsewhere in the card
=> the xpath of the last operation was always wrong.

This was because the normalize function got confused and refused to actually add
the menu template before the card template.

- web_ribbon was appended to the card
=> this caused the style to be broken.

Albeit the ribbon widget makes sense at the top of the card template, it should not break the style
in those proportions.
So, a small fix is done in studio to add the ribbon at the top of the card, but we strongly
recommend a fix that will enable to put it anywhere as a direct child of the relevant view node.

task-4207793

closes odoo/enterprise#71248

X-original-commit: 6d754006f1516491ac503bc0fdef19d9265ac06c
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>