[IMP] web_mobile: lower time_between_reads_in_ms
Lowers the `time_between_reads_in_ms` session setting to 1000 to 100 ms.
The idea is to send more often scanned data.

For now, those data are used only in `stock_barcode` and another batch
is done before processing the data, but it is more suitable to receive
them quicker to update reads count more accurately.

X-original-commit: 6564dca80144ecc20d99e2214d4cc0c381c52ff4
Part-of: odoo/enterprise#70835
Signed-off-by: Steve Van Essche <svs@odoo.com>