[FIX] l10n_ke_edi_oscu_stock: allow negative qty subcontracting location
Steps to reproduce the bug:
- Log in as a Kenyan company
- Create a storable product “P1” with the following Bill of Materials (BoM):
  - Type: Subcontracting
  - Subcontractor: Azure Interior
  - Component:
    - C1:
      - Routes: “Resupply Subcontractor on Order”

- Create a purchase order for one unit of “P1”
- Confirm the PO
- Create a second PO for one unit of “C1” and receive it
- Confirm the resupply picking of C1
- Confirm the receipt of P1

**Problem:**
A validation error is triggered: “You cannot end up with a negative
stock quantity.”

We added a constraint to prevent negative stock quantities in internal
locations for a company in Kenya. However, the location linked to
subcontracting is internal but has no warehouse, so we don’t care if
its quantity is negative, and it should be ignored. This could also
happen with other internal locations that don’t have a warehouse.

Additionally, when reporting stock quantities, we only consider
quantities in locations that have a warehouse:
https://github.com/odoo/enterprise/blob/9dea0f2696084235ee1192b0a14adfa43eb732ef/l10n_ke_edi_oscu_stock/models/product.py#L47

opw-4184857

closes odoo/enterprise#70954

X-original-commit: ef03b2248fda52164e525e5757582cbdc9a7acc2
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>