[IMP] sign: add warning wizard while leaving unsaved document
Before this commit, if a user uploaded a PDF using the "Upload a PDF and sign"
button, added fields to the document, and then left without saving or sending it
for signing, all progress would be lost.

In this commit,  when a user tries to leave a document that hasn't been saved as
a template or sent for signing, a warning dialog will appear. It informs the
user that the document will no longer be available if not saved. The user can
choose to save the document from the dialog if needed, or discard it if they
don't want to keep the changes.

task-4164228

closes odoo/enterprise#71302

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>