[FIX] web_gantt: date fields management
The gantt view was not supposed to support fields of type 'date' but a
view using such fields was introduced in the app "Project" (it is
available in the first menu of the app when the option "Project Stages"
is enabled). Here we make the gantt view manage correctly the date
fields: display, drag and drop, or resize pills should now work as
expected.

Task ID: 3424435

closes odoo/enterprise#71673

X-original-commit: ba1c021cbbeffc3c9af649481ab1751ed67cac2e
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>