[FIX] documents: fix breadcrumbs for non-pinned Company roots
To reproduce:
Open a non-pinned folder in "Company", the breadcrumb will not be correctly
shown and actions are not available.

When simply removed from the returned records, no item is added in the
breadcrumbs on opening a non-pinned company root folder.

Tests are disabled until we fix them with the others.

Task-4216195

closes odoo/enterprise#71176

X-original-commit: d2d5adcc307f949a42d49934a8c420226c6aec36
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Co-authored-by: Lopes Marc (loma) <loma@odoo.com>