[FIX] account_reports: Audit of a custom groupby
Before this commit, auditing a groupby line generated from a custom groupby crashed, because the custom groupby field was assumed to be a field of account.move.line, which it obviously wasn't (that's the very reason custom groupby exist for !).

We solve that by calling the domain builder defined for this groupby in order to generate its audit domain.

This bug is not reproducible in standard Odoo. Custom groupby on reports were implemented to support the 18.0 refactoring of the Intrastat Report, and are at the moment only used there. The Intrastat report does not support audit by clicking the amounts, so it's not subject to the bug. This feature will likely be used more in the future, though.

closes odoo/enterprise#71357

X-original-commit: c9a933fb27f60fa0c13a0bd2c53aed0c17fe7f25
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>