[FIX] planning: copy out-of-schedule shifts as open
Before this commit, when a shift was out of its resource's schedule, we
identified what "half day" it was in and if the resource was working
this half day.
If so, we would assign it, else it would be copied as open.
So say a resource that works from Monday to Friday, 8-12h and 13-17h each
day. If, on the previous week, he had a shift on Tuesday (19-20h), as he
worked this afternoon, he would be assigned to the shift even if it is
outside of his working schedule.
However, since now, the flexible resource exists and has been recently
improved, there is no longer any reason to continue to manage that
specific use case for the resource with working schedule well defined.
That is, now, the shift outside the working schedule should be a open
shift and a flexible resource will take it if he is available.

This commit removes the special behavior for the shifts to copy which are
outside the working schedule of the resource assigned, the new shifts
will be considered as open shifts.

Also, if a fully flexible resource is assigned to a shift, we copy it
no matter what and assign him, since those resources are supposed to be
available.

task-3231732
task-4207839

closes odoo/enterprise#71278

X-original-commit: 0571121bac4a36c307ba7610694b42d0216cdefa
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>