[FIX] account_batch_payment: sign of batch payments
The sign of the batch payment should depend on the direction of the
payment when there are no journal entries, just like when there are
journal entries.
In the case there is one, we were taking the `amount_currency` (which is
direction dependent), but not in the other case.

X-original-commit: 978c71c4b3bb9abbed46f49049cb230c04e937a8
Part-of: odoo/enterprise#71288
Related: odoo/odoo#182470
Related: odoo/upgrade#6586
Signed-off-by: Quentin De Paoli (qdp) <qdp@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>