[IMP] hr_attendance_gantt : improve officer views
Before this commit, when a user is an attendance officer and an attendance manager;
he was able to see some buttons, modify some values but when he tried to
access/save it he had a right issue.

To avoid that, these buttons will be hidden and these fields will be
readonly when the user doesn't have enough attendance right.

task-4044869

closes odoo/enterprise#71349

Related: odoo/odoo#182388
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>