[FIX] industry_fsm: make sure the fsm task cannot be quick created
Before this commit, the changes made in community to call
`get_create_edit_project_ids` method to know which projects we only
allow to quick create a task on `task_id` field in
`account.analytic.line` model (timesheets model) introduced a
performance issue in the list view because the `TaskWithHours` component
doing the rpc call inside `onWillStart` hook is mounted on each row
displayed on the list view, this is, the number of rpc call to that
method will be the number of the records displayed in the list view
instead of calling that rpc once.

This commit moves that behavior in industry_fsm module since
`get_create_edit_project_ids` method will only return fsm projects when
the industry_fsm module is installed. Without that module, no project
ids will be returned and so it is not really needed to do a rpc call for
nothing.
This commit also improves the changes to make sure the rpc call is made
only once in the list view.

Note: that behavior is not managed in the list view of All Timesheets
since it would need to add a custom js class. It will be done in 18.0.

task-4221621

closes odoo/enterprise#71483

X-original-commit: a2af6c085f10dd9443e5337b063054a76ef23dbd
Related: odoo/odoo#182822
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>