[ADD] l10n_se_sie4_import: Import data from SIE 4 files
This commit adds a new module to handle the import of accounting data
from SIE 4 files to Odoo. Odoo can now import accounting data from SIE
of type 4 and below (thanks to backward-compatibility of SIE format).

Even though the SIE 5 is already available in Odoo, the majority of the
Swedish userbase still uses SIE 4 (or below) for importing and exporting
accounting data.

This first implementation does not at all consider the effects of the
`#OBJEKT` and `#DIM` (or other similar dimension) items.

SIE4 files are designed as a text file with label on each line and its
contents. One item can be referred by another item down the file, so we
need a variable (`data_map`) to save the required information for future
use, in-between the reading process.

From the SIE4 file, we can then import the following data: company,
chart of accounts, fiscal year, and move informations.

An option will be provided when creating the wizard to allow user to
choose whether they'd like to import opening balance move, and/or update
any duplicate account data (found in both company and the file)

SIE 4 documentation:
https://sie.se/wp-content/uploads/2020/05/SIE_filformat_ver_4B_ENGLISH.pdf

Task-id: 3735808
X-original-commit: 83f7a6e1c4fec9cc5f5c9c17213c3d0b74cbeddb
Part-of: odoo/enterprise#80647
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Yosua Nicolaus (yoni) <yoni@odoo.com>