[FIX] web_studio: qweb table menu
Since PR odoo/enterprise#181141, when a qweb table menu is displayed, it no longer disappears
when the pointer moves outside it.

This commit will reintroduce the old behavior:
- When the pointer quits a closed qweb table menu, we make it disappear.
- When the pointer leaves an open qweb table menu, it doesn't disappear.
  You need to do a pointerdown to make it disappear.

We'll also add a close to the qweb table menu when scrolling. This will
have the same behavior as the table menu.

closes odoo/enterprise#71249

X-original-commit: 0fb17c40f62ff329998740bb82a6d1138024511c
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Francois Georis (fge) <fge@odoo.com>