[FIX] web_studio: view.normalize: don't aggregate remove and add if different trees
Add a field at the end of the "left" subgroup and remove the "right" subgroup.

```xml

<form>

  <group>

    <group>
      <field name="display_name" />
      [ADD A FIELD]
    </group>
[ REMOVE THIS GROUP
    <group>
    </group>
]
  </group>

</form>
```

Before this commit, when "normalizing" the xpaths, the add field operation was aggregated with the remove operation
because no unchanged node was found in between.
We were left with the inheriting view:
```xml

<xpath expr="[..]/group[2]" position="replace">
  <field name="added_field" />
</xpath>
```

After this commit, this flow works and the two operations are independent.

part of task-4207793

closes odoo/enterprise#72499

X-original-commit: b52e346d36546bf6842fc8c129cb2d36258ebf6f
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>