[FIX] account_accountant: always display other lines in bank reco.
Consider an unreconciled bank statement (line) in the bank
reconciliation widget.
Currently only the liquidity and suspense/transfer line of the
associated move are shown.
This means that some lines of the move are not shown.

After this commit we show all the line (so i.e. the other lines).

Since the diff is not the nicest:
The commit basically just removes the outer `if is_reconciled`.

closes odoo/enterprise#71651

X-original-commit: 68c3baf25e838172d5e4697381e2d82dc096e6ed
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>