[FIX] whatsapp: order of body variables in templates
When users attempt to send a template with 10 body variables, the sample values
for 'Free Text 1' and 'Free Text 10' are swapped.

Steps to Produce:

- Create a template with 10 body variables
- Give sample value for 'Body - {{1}}' as 1 and 'Body - {{10}}' as 10
- Open the WhatsApp composer using the created template
- 'Free text 1' will have a value of 10 instead of 1
- 'Free text 10' will have a value of 1 instead of 10

Problem:

The variables are sorted by name, causing '{{10}}' to appear before '{{1}}',
which leads to the values being swapped.

Solution:

Sort the variables by extracting the index from the variable name, so they are
ordered numerically rather than alphabetically.

Task-4212923

closes odoo/enterprise#81146

X-original-commit: cc792e5452c94b4e1a1a321b2b5e1a4a8b55fc70
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Akshat Trivedi (aktr) <aktr@odoo.com>