[FIX] documents: avoid crashes on inaccessible targets
Reproduce:
1. Create a shortcut for a document you are shared access to.
Say it ends up in My Drive
2. Remove your access to the shortcut's target
3. Go to the folder/section (My Drive) where you have the shortcut
4. Experience an Access Error

Usually, access rights of a document and its shortcuts is synchronized,
so a user can see/edit both or none.
This is not the case with ownership as a user can create a shortcut
 to a document they don't own. If they are allowed to view a shortcut
they own when they don't have access to the target anymore, the webclient
crashes on loading target data (e.g. to show previews).

We tested changing the domain of ownership-related access in
user_permission, but it is very costly performance-wise, so this is
our best approach in stable as it affects writes
but not all reads.

Task-4236268

X-original-commit: a0cf73c1581c49062dac35a957c53d2060448b5a
Part-of: odoo/enterprise#73961
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>