[IMP] hr_recruitment: Improved applicant analysis reports
This PR improves the pivot view of the `Applicant Analysis`, `Source Analysis`, and `Team Performance `
reports. The pivot view for `Source Analysis` and `Team Performance` previously displayed exactly the same
information by default as the `Applicant Analysis` report. This PR improves that by making the default data
of those pivot views the same as what is displayed in the graph view, while also improving the
`Applicant Analysis` report.
- `Applicant analysis` now has `applicant`,`hired`, and `refused` sub columns for each stage.
- `Team Performance` now has the recruiters on one axis and the amount of applicants in each state
  (`in progress`, `hired`, etc.) on the other axis.
- `Source Analysis` now has the recruitment source on one axis and the amount of applicants in each state
  on the other axis.

task-4146657

closes odoo/enterprise#71255

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>