[FIX] l10n_br_avatax: include the newly required barcode for each line
Brazil is now requiring that each line includes the barcode [1]. This
is being rolled out gradually per state, as of now it's only rolled
out in Paraná as far as we are aware.

We're applying this change to Odoo 17 and later because it only
affects EDI. Odoo 16 only supported tax calculation.

Validation is done based on the Avalara documentation [2].

[1] As outlined in "Nota Técnica 2021.003 Validação GTIN"
    https://www.nfe.fazenda.gov.br/portal/exibirArquivo.aspx?conteudo=SrQT9ys8ODo=
[2] https://avataxbr-docs.avalarabrasil.com.br/#/Calculations/payloadCalculation

task-4222168

closes odoo/enterprise#73782

X-original-commit: 6f620c1ed2a0d415f066f8a22573c0d74d1b6a0b
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>