[IMP] account_accountant, account_bank_statement_import: add doc preview
In the bank reconciliation list view, there is no document preview for
the uploaded bank statements. It's not easy to compare transactions with
a statement. This commit remedies that by adding the document preview to
the bank reco list view.

This commit:
- Adds a field bank_statement_attachment_ids and its compute method to
the class AccountBankStatementLine, allowing the link between the line
in the list view and the bank statement attachments.
- Moves the logic from AccountMoveLineListController/Renderer to common
components (AttachmentPreviewListController/Renderer) that both
AccountMoveLineListController/Renderer and BankRecListController/Renderer
inherit from.
- Modifies the BankRecListUploadRenderer to extend
AttachmentPreviewListRenderer instead of ListRenderer
- Modifies the template for the attachment preview in list view by
removing conditions on mailPopoutService.externalWindow, as these
checks were rendered redundant by updates made to the
mail_popout_service (odoo/odoo#174878).
- Removes the override of openRecord in AccountMoveLineListController,
as it is no longer necessary since odoo/odoo#183718
- Removes unecessary use of useState for the mail.store and mail.popout
services

task-4196814

closes odoo/enterprise#71319

Signed-off-by: John Laterre (jol) <jol@odoo.com>