[FIX] appointment: ease booking for new portal users
As new portal users usually have no phone, let us ease
their booking flow while ensuring correct data. Their
phone will be updated if logged and not set.

Encourage them to sign in with a new Sign In button.
Also remove propagation to ecommerce flow so that
we make sure appointment form data is kept as is and
description remains correct.

A test and pot files are adapted accordingly. We check
the number is updated on the logged user's partner if
not set.

Task-4207064

closes odoo/enterprise#73171

X-original-commit: bf709bdb3ce24eab024e2492b9158386bcd585b3
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>