[FIX] spreadsheet_sale_management: add menu
Currently, there's no menu to get a list of all quote calculator spreadsheet
templates.

This commits adds a menu in the technical settings.

It allows to:

- delete/clean unused spreadsheet templates
- import a template from its json file (BA do that quite often for dashboards,
  I expect they'll want to do it for spreadsheet quotation templates)

Feedback from OXP

Note: we don't add the menu in the Sales app menus because we don't want to
add noise in there. It must be kept clean.

Will be backported to 18.0

closes odoo/enterprise#71445

Task: 4236559
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>