[FIX] l10n_mx_edi: Fix global invoice taxes aggregator with round globally
When the global invoice is made on multiple invoices, aggregating rounded values
lead to rounding issues. That's the case for "Subtotal" in the test case added
in this commit:

Suppose 5 invoices having a single invoice line with 16% price included tax. The unit price are:
2803.0, 1842.0, 2798.0, 3225.0, 3371.0.

2803 + 1842 + 2798 + 3225 + 3371 = 14039
14039 / 1.16 = 12102.586206897 ≃ 12102.59
Before this commit, when aggregating each 'Subtotal' of CFDI file, one per invoice:
2416.38 + 1587.93 + 2412.07 + 2780.17 + 2906.03 = 12102.58

X-original-commit: e492ec9da02546deff91737d42e661d28d42c5cd
Part-of: odoo/enterprise#71452
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>