[FIX] spreadsheet_edition: traceback on edit comments
Steps to reproduce: add a comment, edit it, traceback
The override of the composer in composer_patch.js assumed that the props.composer
always had a thread. It is not the case when editing a comment: it has props.composer.message instead

The fix ignores the thread when it isn't there.

Task: 4237047
X-original-commit: 58e67d6ee190ba0e4f4e952df965e0d3fff08cf7
Part-of: odoo/enterprise#71777
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Co-authored-by: alexkuhn