[FIX] l10n_vn_reports: fix traceback when opening the VN tax reports
Currently, a traceback is occurring when the user tries to open a
Vietnamese tax returns report.

To reproduce this issue:

1) Install l10n_vn_reports and switch to VN company
2) Create an invoice with a product and tax
3) Now open the Tax Returns and change the report as
   'Sales/Purchase Tax Report (VN)'

Error:- "ProgrammingError: can't adapt type 'dict'"

For the method `_currency_table_apply_rate` we expect the
value to be in SQL form of string.

For handing the dict for returning join condition we use
`_currency_table_aml_join`.

sentry-5962808992

closes odoo/enterprise#71930

X-original-commit: b682b0a74edc645d98169e08b8f901c572c80253
Signed-off-by: William André (wan) <wan@odoo.com>