[REF] side panel: use title props
Currently, the title of a `Section` component is a title slot. We are using
a slot to translate the title: a props like `title="'My title'"` wouldn't work
since this props is javascript expression, while text content inside xml
elements are extracted and translated.

But since the introduction of the `.translate` modifier, we can use a props!

It's much more natural (props vs slot) and explicit (`.translate`)

Task: 4226704
Part-of: odoo/enterprise#72689
Related: odoo/odoo#185050
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>