[IMP] account_accountant: deferred warning if abnormal duration
In the deferred computations, we always assume that both the start and end date are inclusive
E.g: 1st January -> 31st December is *exactly* 1 year = 12 months
However, the user may instead put 1st January -> 1st January of next year which is then
12 months + 1/30 month = 12.03 months which may result in odd amounts when deferrals are created
For this reason, we now alert the user if we detect such a case with a yellow highlight
in the abnormal dates.
Other cases were the number of months is not round should not be handled.

closes odoo/enterprise#72204

Task-id: none
X-original-commit: cc094e0290f1b5e18c619682f990e5e3d54c09c9
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>